Created views folder. Began useMap.js component
