+++ snarferfrontend/src/components/useMap.js
@@ -1,0 +1,50 @@
+import L from 'leaflet';
+import "leaflet/dist/leaflet.css"; 
+
+export function initializeMap(mapContainerId) {
+  if (!mapContainerId) {
+    console.error("No map container ID provided.");
+    return;
+  }
+
+  // Create map instance
+  const map = L.map(mapContainerId).setView([51.505, -0.09], 13);
+
+  // Add tile layer
+  L.tileLayer('https://tile.openstreetmap.org/{z}/{x}/{y}.png', {
+    maxZoom: 19,
+    attribution: '&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a>'
+  }).addTo(map);
+
+  // Add objects to the map
+  const marker = L.marker([51.5, -0.09]).addTo(map);
+  const circle = L.circle([51.508, -0.11], {
+    color: 'red',
+    fillColor: '#f03',
+    fillOpacity: 0.5,
+    radius: 500
+  }).addTo(map);
+
+  const polygon = L.polygon([
+    [51.509, -0.08],
+    [51.503, -0.06],
+    [51.51, -0.047]
+  ]).addTo(map);
+
+  // Bind popups
+  marker.bindPopup("<b>Hello world!</b><br>I am a popup.").openPopup();
+  circle.bindPopup("I am a circle.");
+  polygon.bindPopup("I am a polygon.");
+
+  // Click event
+  const popup = L.popup();
+  function onMapClick(e) {
+    popup
+      .setLatLng(e.latlng)
+      .setContent("You clicked the map at " + e.latlng.toString())
+      .openOn(map);
+  }
+  map.on('click', onMapClick);
+
+  return map;
+}
