+++ snarferfrontend/src/main.js
@@ -4,11 +4,11 @@ import { createApp } from 'vue'
 import { createPinia } from 'pinia'
 import { createRouter, createWebHistory } from 'vue-router'
 import App from './App.vue'
-import router from './router' // Import the router
+import router from './router'
 
 const app = createApp(App)
 
 app.use(createPinia())
-app.use(router) // Use the router
+app.use(router)  // Temporary? Should we use vue-router instead?
 
 app.mount('#app')
