+++ snarferfrontend/src/router.js
@@ -1,16 +1,16 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import LoginPage from './components/LoginPage.vue'
-import Splashy from './components/Splashy.vue'
-import createAccount from './components/createAccount.vue'
-import homePage from './components/homePage.vue'
-import catEdit from './components/catEdit.vue'
+import LoginPage from './views/LoginPage.vue'
+import Splashy from './views/Splashy.vue'
+import createAccount from './views/createAccount.vue'
+import homePage from './views/homePage.vue'
+import catEdit from './views/catEdit.vue'
 
 const routes = [
   { path: '/', component: Splashy },  // Default page (Landing)
   { path: '/login', component: LoginPage },
   { path: '/createAccount', component: createAccount},
-  { path: '/homePage', component: homePage},
-  { path: '/catEdit', component: catEdit},
+  { path: '/homePage', component: homePage},   //Main page for logged in users.
+  { path: '/catEdit', component: catEdit},  //Page for editing Catalog entries
   { path: '/:pathMatch(.*)*', redirect: '/' } // Redirect all other paths to Splashy
 ]
