+++ snarferfrontend/src/views/LoginPage.vue
