+++ snarferfrontend/src/views/Splashy.vue
@@ -1,6 +1,6 @@
 <script setup>
-import HelloWorld from './HelloWorld.vue'
-import TheWelcome from './TheWelcome.vue'
+import HelloWorld from '@/components/HelloWorld.vue'
+import TheWelcome from '@/components/TheWelcome.vue'
 </script>
 
 <template>
