+++ snarferfrontend/src/views/catEdit.vue
