+++ snarferfrontend/src/views/createAccount.vue
